Move getAliases to SelectStatement
