+++ psalm-baseline.xml
@@ -5177,6 +5177,9 @@
     </PossiblyNullOperand>
   </file>
   <file src="src/Statements/SelectStatement.php">
+    <PossiblyNullArrayOffset>
+      <code>$tables[$thisDb]</code>
+    </PossiblyNullArrayOffset>
     <PossiblyUnusedProperty>
       <code>$endOptions</code>
       <code>$groupOptions</code>
@@ -5418,11 +5421,6 @@
     <RedundantConditionGivenDocblockType>
       <code><![CDATA[$curr->type === Token::TYPE_KEYWORD]]></code>
     </RedundantConditionGivenDocblockType>
-  </file>
-  <file src="src/Utils/Misc.php">
-    <PossiblyNullArrayOffset>
-      <code>$tables[$thisDb]</code>
-    </PossiblyNullArrayOffset>
   </file>
   <file src="src/Utils/Query.php">
     <InvalidNullableReturnType>
